TS migration: Entry, Protein, Proteome, Structure
