+++ src/pages/index.js
@@ -36,9 +36,11 @@ const Entry = loadable({
   loader: () => import(/* webpackChunkName: "entry-page" */ './Entry'),
 });
 const Protein = loadable({
+  // $FlowFixMe
   loader: () => import(/* webpackChunkName: "protein-page" */ './Protein'),
 });
 const Structure = loadable({
+  // $FlowFixMe
   loader: () => import(/* webpackChunkName: "structure-page" */ './Structure'),
 });
 const Taxonomy = loadable({
